pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "event-platform"
version = "0.1.0"
description = "High-performance event & marketing platform (Django-based)."
readme = "README.md"
requires-python = ">=3.12,<3.13"
authors = [
    { name = "Mats Gustafsson", email = "mats.gustafsson83@gmail.com" },
]
dependencies = [
    "Django>=5.1,<6.0",
    "djangorestframework>=3.15,<4.0",
    "djangorestframework-simplejwt>=5.3,<6.0",
    "django-htmx>=1.21,<2.0",
    "django-redis>=5.4,<6.0",
    "psycopg[binary,pool]>=3.2,<4.0",
    "python-decouple>=3.8,<4.0",
    "Pillow>=11.0,<12.0",
    "whitenoise>=6.7,<7.0",
    "gunicorn>=23.0,<24.0",
    "uvicorn[standard]>=0.30,<1.0",
    "cloudinary>=1.41,<2.0",
    "django-cloudinary-storage>=0.3.0,<0.4.0",
    "django-cors-headers>=4.4,<5.0",
    "wagtail>=6.0,<8.0",
]


[project.optional-dependencies]
dev = [
    "ruff>=0.7,<0.8",
    "mypy>=1.11,<2.0",
    "pytest>=8.3,<9.0",
    "pytest-django>=4.9,<5.0",
    "coverage[toml]>=7.6,<8.0",
    "pre-commit>=4.0,<5.0",
    "django-stubs[compatible-mypy]>=5.0,<6.0",
    "djangorestframework-stubs>=3.15,<4.0",
    "types-requests>=2.32,<3.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = []

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
ignore_missing_imports = false
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]

[[tool.mypy.overrides]]
module = "decouple"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "config.settings"
ignore_errors = true

[[tool.mypy.overrides]]
module = "wagtail.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "modelcluster.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "taggit.*"
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py"]
addopts = "-ra -q --disable-warnings"

[tool.coverage.run]
branch = true
source = ["apps", "config"]

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["config*", "apps*"]
exclude = ["static*", "templates*"]
